Move commands to the adapters folder
