+++ apps/cli/src/adapters/commander/commands/doctor.ts
@@ -1,8 +1,8 @@
 import { Command } from "commander";
 import * as process from "node:process";
 
-import { Dependencies } from "../domain/dependencies.js";
-import { checkMonorepoScripts } from "../domain/node.js";
+import { Dependencies } from "../../../domain/dependencies.js";
+import { checkMonorepoScripts } from "../../../domain/node.js";
 
 type DoctorDependencies = Pick<
   Dependencies,
