+++ apps/cli/src/adapters/commander/commands/index.ts
